moved zero tolerance and led constants
