+++ src/main/java/frc/robot/constants/Constants.java
@@ -17,7 +17,6 @@ public final class Constants {
     public static final int PRIMARY_XBOX_CONTROLLER_PORT = 0;
     public static final String CANIVORE_BUS = "canivoreBus";
     public static final String RIO_BUS = "rio";
-    public static final double ZERO_TOLERANCE = 0.005;
     public static final int SECONDARY_XBOX_CONTROLLER_PORT = 1;
     public static final int GIGA_PORT = 3;
 }
