+++ src/main/java/frc/robot/subsystems/led/LEDConstants.java
@@ -1,4 +1,4 @@
-package frc.robot.constants;
+package frc.robot.subsystems.led;
 
 import static edu.wpi.first.units.Units.Meters;
 
@@ -13,6 +13,7 @@ public final class LEDConstants {
     public static final Color YETI_BLUE = new Color(84, 182, 229);
     public static final Color ALGAE_COLOR = new Color(79, 181, 165);
     public static final Color NICK_ORANGE = new Color(240, 119, 99);
+    public static final double ZERO_TOLERANCE = 0.005;
     private static final int PROGRESS_PARTS = 4;
     public static final double PROGRESS_INCREMENT = 1.0 / PROGRESS_PARTS;
 }
